Improve how Q is extracted from qrfact!
The previous code was very inefficient because it seemed that
    getindex(A::Union{QRPackedQ,QRCompactWYQ}, i::Integer, j::Integer)
(in qr.jl) was repeatedly called with all combinations of i and j.

Reading the official document
https://docs.julialang.org/en/latest/stdlib/LinearAlgebra/#LinearAlgebra.qrfact
it seems that Matrix(F[:Q]) is the "official" way to get it.  However,
microbenchmarks showed that A_mul_B!(F[:Q], Q) was equally/slightly
faster.  It also seemed that A_mul_B! was slightly better when Q was
"thin" rectangular matrix (i.e., size(Q, 1) >> size(Q, 2)) due to
copying overhead.